microservices/upload_repo_on_blob/requirements.txt
fastapi
uvicorn
python-dotenv
azure-storage-blob
